==> 3.combine_results.ipynb <==
# Notebook to combine results from all experiments
## 20/12/2021
### Required python version: 3.6 

The following is a notebook that will combine the results from all the different date organised folders into one master spreadsheet. 
The results were generated using the batch_biomechanics_csv.py script in combination with the run_analysis.sh script:

```

for i in $(cut -f 1 date_index.tsv); do
   echo ${i}
   cd ${i}

   cp /mnt/share/EMILYJ-CompMod/biomechanics_organised_python_analysis/tendon_data_formatted.csv ./
   python /mnt/share/EMILYJ-CompMod/biomechanics_organised_python_analysis/batch_biomechanics_csv.py

   cd ..

done
```
<br>Contact Emily Johnson at ejohn16@liv.ac.uk or em.j.johnson.93@gmail.com if you're having trouble getting this notebook to work. 

# Load packages

import pandas as pd
import numpy as np
import os
import sys 
import re

dir = os.getcwd()
subfolders = [ f.name for f in os.scandir(dir) if f.is_dir() ]

summary_data = {}
for folder in subfolders:
    summary_data[folder] = pd.read_csv("{}/{}/results_summary.csv".format(dir, folder), header=0)

all_data = pd.concat(summary_data.values(), ignore_index=True)

all_data.to_csv("{}/all_results_summary.csv".format(dir), index=False)
